requirements.txt
torch>=2.0.0
torchvision
torchxrayvision
numpy
pillow
matplotlib
python-dotenv
google-generativeai
